+++ internal/routing/mock.go
@@ -2,17 +2,20 @@ package routing
 
 import (
 	"context"
+	"net/netip"
 	"sync"
 )
 
 type MockRouter struct {
-	resolver map[string][]string
+	resolver map[string][]netip.AddrPort
+	self     netip.AddrPort
 	mx       sync.RWMutex
 }
 
-func NewMockRouter(resolver map[string][]string) *MockRouter {
+func NewMockRouter(resolver map[string][]netip.AddrPort, self netip.AddrPort) *MockRouter {
 	return &MockRouter{
 		resolver: resolver,
+		self:     self,
 	}
 }
 
@@ -26,8 +29,8 @@ func (m *MockRouter) HasMirrors() (bool, error) {
 	return len(m.resolver) > 0, nil
 }
 
-func (m *MockRouter) Resolve(ctx context.Context, key string, allowSelf bool, count int) (<-chan string, error) {
-	peerCh := make(chan string, count)
+func (m *MockRouter) Resolve(ctx context.Context, key string, allowSelf bool, count int) (<-chan netip.AddrPort, error) {
+	peerCh := make(chan netip.AddrPort, count)
 	peers, ok := m.resolver[key]
 	// Not found will look forever until timeout.
 	if !ok {
@@ -48,12 +51,12 @@ func (m *MockRouter) Advertise(ctx context.Context, keys []string) error {
 	m.mx.Lock()
 	defer m.mx.Unlock()
 	for _, key := range keys {
-		m.resolver[key] = []string{"localhost"}
+		m.resolver[key] = []netip.AddrPort{m.self}
 	}
 	return nil
 }
 
-func (m *MockRouter) LookupKey(key string) ([]string, bool) {
+func (m *MockRouter) LookupKey(key string) ([]netip.AddrPort, bool) {
 	m.mx.RLock()
 	defer m.mx.RUnlock()
 	v, ok := m.resolver[key]