+++ internal/routing/routing.go
@@ -2,6 +2,7 @@ package routing
 
 import (
 	"context"
+	"net/netip"
 	"time"
 )
 
@@ -9,7 +10,7 @@ const KeyTTL = 10 * time.Minute
 
 type Router interface {
 	Close() error
-	Resolve(ctx context.Context, key string, allowSelf bool, count int) (<-chan string, error)
+	Resolve(ctx context.Context, key string, allowSelf bool, count int) (<-chan netip.AddrPort, error)
 	Advertise(ctx context.Context, keys []string) error
 	HasMirrors() (bool, error)
 }